+++ packages/modular-scripts/src/__tests__/test.test.ts
@@ -8,6 +8,7 @@ import {
   runModularPipeLogs,
 } from '../test/utils';
 import getModularRoot from '../utils/getModularRoot';
+import { runTestForTests, setupMocks } from '../test/mockFunctions';
 
 const modularRoot = getModularRoot();
 
@@ -17,6 +18,7 @@ describe('Modular test command', () => {
 
     beforeEach(() => {
       tempModularRepo = createModularTestContext();
+      setupMocks(tempModularRepo);
       const fixturesFolder = path.join(__dirname, '__fixtures__', 'test');
       const relativeFixturePath = fixturesFolder.replace(modularRoot, '');
       const tempFixturesFolder = path.join(
@@ -36,32 +38,21 @@ describe('Modular test command', () => {
     });
 
     describe('when the tests fail', () => {
-      it('should exit with an error', async () => {
-        let errorNumber = 0;
-        try {
-          await runYarnModular(
-            tempModularRepo,
-            'test --regex test/InvalidTest.test.ts --watchAll=false',
-          );
-        } catch (error) {
-          errorNumber = (error as ExecaError).exitCode;
-        }
-        expect(errorNumber).toBe(1);
+      it('should error', async () => {
+        const { err } = await runTestForTests({
+          regex: ['test/InvalidTest.test.ts'],
+        });
+        expect(err).toContain('Modular test did not pass');
       });
     });
 
     describe('when the tests pass', () => {
-      it('should exit with no error', async () => {
-        let errorNumber = 0;
-        try {
-          await runYarnModular(
-            tempModularRepo,
-            'test --regex test/ValidTest.test.ts --watchAll=false',
-          );
-        } catch (error) {
-          errorNumber = (error as ExecaError).exitCode;
-        }
-        expect(errorNumber).toBe(0);
+      it('should not error', async () => {
+        const { mockExit, err } = await runTestForTests({
+          regex: ['test/ValidTest.test.ts'],
+        });
+        expect(err).toBeUndefined();
+        expect(mockExit).not.toHaveBeenCalled();
       });
     });
   });
@@ -79,7 +70,7 @@ describe('Modular test command', () => {
 
     let randomOutputFolder: string;
 
-    beforeEach(() => {
+    beforeAll(() => {
       // Create random dir
       randomOutputFolder = tmp.dirSync({ unsafeCleanup: true }).name;
       fs.copySync(fixturesFolder, randomOutputFolder);
@@ -113,17 +104,16 @@ describe('Modular test command', () => {
       });
     });
 
-    // These expects run in a single test, serially for performance reasons (the setup time is quite long)
-    it('finds no unchanged using --changed / finds changed after modifying some workspaces / finds ancestors using --ancestors', () => {
-      const resultUnchanged = runModularPipeLogs(
+    it('finds no unchanged using --changed', () => {
+      const result = runModularPipeLogs(
         randomOutputFolder,
         'test --changed',
         'true',
       );
-      expect(resultUnchanged.stdout).toContain(
-        'No workspaces found in selection',
-      );
+      expect(result.stdout).toContain('No workspaces found in selection');
+    });
 
+    it('finds changed after modifying some workspaces', () => {
       fs.appendFileSync(
         path.join(randomOutputFolder, '/packages/b/src/index.ts'),
         "\n// Comment to package b's source",
@@ -133,28 +123,49 @@ describe('Modular test command', () => {
         "\n// Comment to package c's source",
       );
 
-      const resultChanged = runModularPipeLogs(
+      const result = runModularPipeLogs(
         randomOutputFolder,
         'test --changed',
         'true',
       );
-      expect(resultChanged.stderr).toContain('c-nested.test.ts');
-      expect(resultChanged.stderr).toContain('c.test.ts');
-      expect(resultChanged.stderr).toContain('b-nested.test.ts');
-      expect(resultChanged.stderr).toContain('b.test.ts');
+      expect(result.stderr).toContain('c-nested.test.ts');
+      expect(result.stderr).toContain('c.test.ts');
+      expect(result.stderr).toContain('b-nested.test.ts');
+      expect(result.stderr).toContain('b.test.ts');
+    });
 
-      const resultChangedWithAncestors = runModularPipeLogs(
+    it('finds ancestors using --ancestors', () => {
+      const result = runModularPipeLogs(
         randomOutputFolder,
         'test --changed --ancestors',
       );
-      expect(resultChangedWithAncestors.stderr).toContain('c-nested.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('c.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('b-nested.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('b.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('a-nested.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('a.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('e-nested.test.ts');
-      expect(resultChangedWithAncestors.stderr).toContain('e.test.ts');
+      expect(result.stderr).toContain('c-nested.test.ts');
+      expect(result.stderr).toContain('c.test.ts');
+      expect(result.stderr).toContain('b-nested.test.ts');
+      expect(result.stderr).toContain('b.test.ts');
+      expect(result.stderr).toContain('a-nested.test.ts');
+      expect(result.stderr).toContain('a.test.ts');
+      expect(result.stderr).toContain('e-nested.test.ts');
+      expect(result.stderr).toContain('e.test.ts');
+    });
+
+    it('correctly interprets Jest options when no selective options are provided', () => {
+      const result = runModularPipeLogs(
+        randomOutputFolder,
+        'test --testNamePattern="should add two numbers unique"',
+      );
+      expect(result.stderr).toContain(
+        `Ran all test suites matching /packages\\/a|packages\\/b|packages\\/c|packages\\/d|packages\\/e/i with tests matching "should add two numbers unique".`,
+      );
+    });
+    it('correctly interprets Jest options when selective options are provided', () => {
+      const result = runModularPipeLogs(
+        randomOutputFolder,
+        'test a --testNamePattern="should add two numbers unique"',
+      );
+      expect(result.stderr).toContain(
+        'Ran all test suites matching /packages\\/a/i with tests matching "should add two numbers unique"',
+      );
     });
   });