+++ packages/modular-scripts/src/lint.ts
@@ -11,6 +11,10 @@ import * as logger from './utils/logger';
 import { generateJestConfig } from './test/utils';
 import { getAllWorkspaces } from './utils/getAllWorkspaces';
 import { selectWorkspaces } from './utils/selectWorkspaces';
+import {
+  computeRegexesFromPackageNames,
+  partitionPackages,
+} from './utils/unobtrusiveModular';
 export interface LintOptions {
   all: boolean;
   fix: boolean;
@@ -19,12 +23,13 @@ export interface LintOptions {
   ancestors?: boolean;
   descendants?: boolean;
   changed?: boolean;
+  includeNonModular?: boolean;
   compareBranch?: string;
 }
 
 async function lint(
   options: LintOptions,
-  regexes: string[] = [],
+  userRegexes: string[] = [],
 ): Promise<void> {
   const {
     all = false,
@@ -34,57 +39,65 @@ async function lint(
     ancestors = false,
     descendants = false,
     changed = false,
+    includeNonModular = false,
     compareBranch,
   } = options;
   const modularRoot = getModularRoot();
   const lintExtensions = ['.ts', '.tsx', '.js', '.jsx'];
-  let runnerMatch = ['<rootDir>/**/src/**/*.{js,jsx,ts,tsx}'];
 
-  const [packageMap] = await getAllWorkspaces();
-
-  const selectiveOptionSpecified =
+  const [workspaceMap] = await getAllWorkspaces();
+  const isSelective =
     packages?.length || ancestors || descendants || changed || compareBranch;
 
-  if (all && selectiveOptionSpecified) {
+  let selectedTargets: string[];
+
+  if (!isSelective) {
+    selectedTargets = [...workspaceMap.keys()];
+  } else {
+    // Otherwise, calculate which packages are selected
+    selectedTargets = await selectWorkspaces({
+      targets: packages,
+      changed,
+      ancestors,
+      descendants,
+      compareBranch,
+    });
+  }
+
+  // Split packages into modular and non-modular testable. Make sure that "root" is not there.
+  const [modularTargets, nonModularTargets] = partitionPackages(
+    selectedTargets,
+    workspaceMap,
+    'lint',
+  );
+
+  // Compute patterns to pass Jest for the packages that we want to test
+  const selectedPackageRegexes = await computeRegexesFromPackageNames(
+    modularTargets,
+  );
+
+  // Only set regexes if all or selective options have been specified
+  let processedPackageRegexes = selectedPackageRegexes.map(
+    (regex) => `<rootDir>/${regex}/src/**/*.{js,jsx,ts,tsx}`,
+  );
+
+  let lintRegexes = isSelective || all ? processedPackageRegexes : [];
+
+  if (all && isSelective) {
     logger.warn(
       'You specified --all already; selective options are redundant.',
     );
   }
-  if (staged && selectiveOptionSpecified) {
+  if (staged && isSelective) {
     logger.error("Can't specify --staged along with selective options");
     process.exit(-1);
   }
 
-  if (!all) {
-    if (selectiveOptionSpecified) {
-      // If at least one of the selective options is specified, calculate target packages from selective options
-      const targetPackages = await selectWorkspaces({
-        targets: packages,
-        changed,
-        ancestors,
-        descendants,
-        compareBranch,
-      });
-
-      // Calculate the match regexes for the selective options
-      const selectiveMatch = targetPackages
-        .map((packageName) => {
-          const packageLocation = packageMap.get(packageName)?.location;
-          return packageLocation
-            ? `<rootDir>/${packageLocation}/src/**/*.{js,jsx,ts,tsx}`
-            : undefined;
-        })
-        .filter(Boolean) as string[];
-
-      // If there are no selective matches and no regexes are specified, there is nothing to test
-      if (!selectiveMatch.length && !regexes.length) {
-        logger.warn('No target files to lint with the provided selection');
-        return;
-      }
-      // Narrow the matches to the selection matches
-      runnerMatch = selectiveMatch;
-    } else if ((!isCI || staged) && regexes.length === 0) {
-      // Not selective, not --all and no regexes; calculate the file regexes of --diff or --staged
+  // Not selective, not --all and no regexes;
+  // TODO: Bring this in line with behaviour of other commands, too 'unpredictable' from a user point of view
+  if (!all && !isSelective && userRegexes.length === 0) {
+    if (staged || !isCI) {
+      // If --staged or not in CI, calculate the file regexes of --diff or --staged
       let diffedFiles: null | string[];
 
       try {
@@ -95,65 +108,106 @@ async function lint(
         logger.log(
           'Getting staged or diffed files failed - are you sure this is a git repo? Falling back to `--all`.',
         );
+        lintRegexes = processedPackageRegexes;
       }
-
       if (diffedFiles !== null) {
         if (diffedFiles.length === 0) {
           logger.log(
             'No diffed files detected. Use the `--all` option to lint the entire codebase',
           );
-          return;
-        }
-        const targetExts = diffedFiles
-          .filter((p: string) => lintExtensions.includes(path.extname(p)))
-          .map((p: string) => `<rootDir>/${p}`);
+        } else {
+          processedPackageRegexes = diffedFiles
+            .filter((p: string) => lintExtensions.includes(path.extname(p)))
+            .map((p: string) => `<rootDir>/${p}`);
 
-        // if none of the diffed files meet the extension criteria, do not lint
-        // end the process early with a success
-        if (!targetExts.length) {
-          logger.warn('No diffed target files to lint found');
-          return;
+          // if none of the diffed files meet the extension criteria, do not lint
+          // end the process early with a success
+          if (!processedPackageRegexes.length) {
+            logger.warn('No diffed target files to lint found');
+          }
         }
-        runnerMatch = targetExts;
       }
+    } else {
+      // If in CI, then lint all
+      lintRegexes = processedPackageRegexes;
     }
+  }
+
+  // If we computed no regexes and there are no non-modular packages to lint, bail out
+  if (!lintRegexes?.length && !userRegexes && !nonModularTargets.length) {
+    process.stdout.write('No workspaces found in selection\n');
+    process.exit(0);
   }
 
   const jestEslintConfig = {
     runner: require.resolve('modular-scripts/jest-runner-eslint'),
     displayName: 'lint',
     rootDir: modularRoot,
-    testMatch: runnerMatch,
+    testMatch: lintRegexes,
     testPathIgnorePatterns: ['/node_modules/', '/dist/'],
   };
 
-  const testArgs = [
-    ...regexes,
+  const lintArgs = [
+    ...userRegexes,
     '--config',
     generateJestConfig(jestEslintConfig),
   ];
 
-  const testBin = await resolveAsBin('jest-cli');
+  const lintBin = await resolveAsBin('jest-cli');
 
-  try {
-    await execAsync(testBin, testArgs, {
-      cwd: modularRoot,
-      log: false,
-      // @ts-ignore
-      env: {
-        MODULAR_ROOT: modularRoot,
-        MODULAR_LINT_FIX: String(fix),
-      },
-    });
+  logger.debug(
+    `Regexes for Modular packages being linted: ${JSON.stringify(lintRegexes)}`,
+  );
+  logger.debug(`User provided regexes: ${JSON.stringify(userRegexes)}`);
 
-    if (staged && fix) {
-      runnerMatch = runnerMatch.map((p) => p.replace('<rootDir>/', ''));
-      addFiles(runnerMatch);
+  // Lint modular packages
+  if (lintRegexes.length || userRegexes.length) {
+    try {
+      logger.debug(
+        `Running ${lintBin} with cwd: ${getModularRoot()} and args: ${JSON.stringify(
+          lintArgs,
+        )}`,
+      );
+      await execAsync(lintBin, lintArgs, {
+        cwd: modularRoot,
+        log: false,
+        // @ts-ignore
+        env: {
+          MODULAR_ROOT: modularRoot,
+          MODULAR_LINT_FIX: String(fix),
+        },
+      });
+
+      if (staged && fix) {
+        lintRegexes = lintRegexes.map((p) => p.replace('<rootDir>/', ''));
+        addFiles(lintRegexes);
+      }
+    } catch (err) {
+      logger.debug((err as ExecaError).message);
+      // ✕ Modular lint did not pass
+      throw new Error('\u2715 Modular lint did not pass');
     }
-  } catch (err) {
-    logger.debug((err as ExecaError).message);
-    // ✕ Modular lint did not pass
-    throw new Error('\u2715 Modular lint did not pass');
+  }
+
+  // Lint non-modular packages - TODO: Remove conditional & flag in next major release of Modular (5.0.0)
+  if (includeNonModular) {
+    try {
+      logger.debug(
+        `Running lint command in the following non-modular packages: ${JSON.stringify(
+          nonModularTargets,
+        )}`,
+      );
+      for (const target of nonModularTargets) {
+        await execAsync(`yarn`, ['workspace', target, 'lint'], {
+          cwd: getModularRoot(),
+          log: false,
+        });
+      }
+    } catch (err) {
+      logger.debug((err as ExecaError).message);
+      // ✕ Modular lint did not pass
+      throw new Error('\u2715 Modular lint did not pass');
+    }
   }
 }