+++ packages/modular-scripts/src/utils/selectWorkspaces.ts
@@ -46,7 +46,7 @@ export async function selectWorkspaces(
 }
 
 /**
- * Select buildable target packages in workspaces in build order, optionally including changed, ancestors and descendant packages
+ * Select buildable target packages in build order, optionally including changed, ancestors and descendant packages. The result is returned as a 1-level nested array of package names, where sub-arrays can be built parallely
  *
  * Please note that the build order algorithm can't calculate build order if there's a cycle in the dependency graph,
  * so circular dependencies will throw unless they only involve packages that don't get built (i.e. "source" `modular.type`s).
@@ -55,12 +55,12 @@ export async function selectWorkspaces(
  * and they make your code fragile towards refactoring, so please don't introduce them in your monorepository.
  *
  * @param  {SelectBuildableOptions} options The target options to configure selection
- * @return {Promise<string[]>} A distinct list of selected buildable package names, in build order
+ * @return {Promise<string[][]>} A nested (1 level depth) list of selected buildable package names, in build order. Nested subarrays can be built parallely.
  */
 
-export async function selectBuildableWorkspaces(
+export async function selectParallellyBuildableWorkspaces(
   options: SelectBuildableOptions,
-): Promise<string[]> {
+): Promise<string[][]> {
   const { ancestors, descendants, dangerouslyIgnoreCircularDependencies } =
     options;
   const {
@@ -113,22 +113,30 @@ export async function selectBuildableWorkspaces(
 
   return (
     targetEntriesWithOrder
-      .sort((a, b) => b[1] - a[1])
-      .map(([packageName]) => packageName)
       // Filter out descendants and ancestors if we don't explicitly need them.
       // This allows us to get the correct dependency order even if we restrict the scope (for example, by explicit user input).
       .filter(
-        (packageName) =>
+        ([packageName]) =>
           (descendants && descendantsSet.has(packageName)) ||
           (ancestors && ancestorsSet.has(packageName)) ||
           targetsToBuild.includes(packageName),
       )
+      // The output is a sequence of [packageName, level] - transform it into a sequence of [[level_0_pkg, ...], [level_1_pkg, ...], ...]
+      .reduce<string[][]>((acc, [pkg, level]) => {
+        acc[level] ? acc[level].push(pkg) : (acc[level] = [pkg]);
+        return acc;
+      }, [])
+      // The previous reduction can leave holes in the array, let's remove them.
+      .filter(Boolean)
+      // Reverse in actual build order
+      .reverse()
   );
 }
 
 /**
- * Common data structures to calculate package selection
+ * Common data structures / functions to calculate package selection
  */
+
 async function computeWorkspaceSelection({
   targets,
   changed,